pyproject.toml
[tool.poetry]
name = "onlinetraining"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
